App becomes lifecycleware . Old chats are listing, User can start new chat.

+++ app/src/main/java/com/papara/geminiapp/data/local/entity/ChatMessage.kt
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "chat_messages")
 data class ChatMessage(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val conversationId: Int,
+    val conversationId: Long,
     val message: String,
     val isFromUser: Boolean,
     val timestamp: Long=System.currentTimeMillis()
